added more sensor data and improved madgwick filter accuracy and speed

+++ AHRS-dev/src/FusionAhrs.h
@@ -80,9 +80,9 @@ void FusionAhrsReset(FusionAhrs* const ahrs);
 
 void FusionAhrsSetSettings(FusionAhrs* const ahrs, const FusionAhrsSettings* const settings);
 
-void FusionAhrsUpdate(FusionAhrs* const ahrs, const FusionVector gyroscope, const FusionVector accelerometer, const FusionVector magnetometer, const float deltaTime);
+void FusionAhrsUpdate(FusionAhrs* const ahrs, const FusionVector gyroscope, const FusionVector accelerometer, const FusionVector magnetometer, const double deltaTime);
 
-void FusionAhrsUpdateNoMagnetometer(FusionAhrs* const ahrs, const FusionVector gyroscope, const FusionVector accelerometer, const float deltaTime);
+void FusionAhrsUpdateNoMagnetometer(FusionAhrs* const ahrs, const FusionVector gyroscope, const FusionVector accelerometer, const double deltaTime);
 
 void FusionAhrsUpdateExternalHeading(FusionAhrs* const ahrs, const FusionVector gyroscope, const FusionVector accelerometer, const float heading, const float deltaTime);
